made slight change to scping tocrack to pis & added check for client ip to distribute john

+++ Scripts/distribute_john_binary.sh
@@ -4,9 +4,11 @@ cd ~/
 tar cvzf john.tar.gz JohnTheRipper
 for i in $(nmap -sP 192.168.1.30-45 | grep -o -P "\d+\.\d+\.\d+.\d+");
 do
-	scp john.tar.gz pi@$i:~/
-	ssh pi@$i "rm -rf JohnTheRipper"
-	ssh pi@$i "tar -xf john.tar.gz"
-	echo "Test $i"
-	ssh pi@$i "./JohnTheRipper/run/john -t -fo:odf"
+	if [ "$1" != "$i" ] ; then
+		scp john.tar.gz pi@$i:~/
+		ssh pi@$i "rm -rf JohnTheRipper"
+		ssh pi@$i "tar -xf john.tar.gz"
+		echo "Test $i"
+		ssh pi@$i "./JohnTheRipper/run/john -t -fo:odf"
+	fi
 done
